% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jexcel.R
\name{excel}
\alias{excel}
\title{Create excel table using jexcel library}
\usage{
excel(data = NULL, columns = NULL, colHeaders = NULL,
  rowHeight = NULL, nestedHeaders = NULL, minDimensions = NULL,
  search = FALSE, pagination = NULL, allowComments = FALSE,
  mergeCells = NULL, columnSorting = FALSE, lazyLoading = TRUE,
  loadingSpin = TRUE, style = NULL, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{data}{a data object, can either be dataframe or a matrix}

\item{columns}{a datframe containing the attributes of columns in the table.
The row number  of the dataframe specifies the column for with the attribute is to be specified
Each header of the dataframe specifies the attribute.The following attributes are supported:
 \itemize{
  \item \strong{title:} the title of the column or the column header
  \item \strong{width:} the width of the column
  \item \strong{type:} the type of the column. It can be one of text, numeric, hidden, dropdown,
  autocomplete, checkbox, radio, calendar, image or color
  \item \strong{source:} the options for column whose type is 'dropdwown'.
   We need to specify the source if type of column is 'dropdown'
   \item \strong{multiple:} used to specify if the multiple options can be selected if the type is 'dropdown'.
   The default value is false.
}}

\item{colHeaders}{the column header, vector of column headers. If both colHeader and title attribute
in columns is specified, the latter will take precedence}

\item{rowHeight}{a dataframe or matrix specifying heights of different rows. The first column specifies
the row number while the second column speicifies the height in pixels}

\item{nestedHeaders}{a list of dataframe having title and colspan as the attributes. The nested header
in the same level should be in the same dataframe}

\item{minDimensions}{a vector that defines the minimum dimension size of the table irrespective of the size of data,
the first parameter should be of column followed by row. If data is null then it will create an empty table with the given
dimensions}

\item{search}{a boolean value indicating if search should be enabled.}
}
\description{
This function is used to create excel like table
}
