% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selected_data.R
\name{get_selected_data}
\alias{get_selected_data}
\title{Convert excel object to data.frame}
\usage{
get_selected_data(excelObj)
}
\arguments{
\item{excelObj}{the json data retuned from excel table}
}
\description{
This function is used to the data selected in jexcel table
}
\examples{
if(interactive()){
 library(shiny)
 library(excelR)
 shinyApp(
   ui = fluidPage(excelOutput("table")),
   server = function(input, output, session) {
     output$table <-
       renderExcel(excelTable(data = head(iris)))
     observeEvent(input$table,{
       print(get_selected_data(input$table))
     })
   }
 )
}
}
