% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_data.R
\name{getTableData}
\alias{getTableData}
\title{This function is used to get data from excel table}
\usage{
getTableData(tableId)
}
\arguments{
\item{tableId}{the id of the table from which the comment is to be fetched}
}
\description{
This function is used to get data from excel table
}
\examples{
if(interactive()) {
 library(excelR)
 library(shiny)
     shinyApp(
         ui = fluidPage(actionButton('get', 'Get table data'),
             tableOutput("fetchedData"),
             excelOutput("table", height = 175)),
         server = function(input, output, session) {
             output$table <- renderExcel(excelTable(data = head(iris)))

             observeEvent(input$get,{getTableData("table")})

             observeEvent(input$table,{
                 output$fetchedData <- renderTable(excel_to_R(input$table))
             })
})
}

}
