% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide_show_index_col.R
\name{hideShowIndex}
\alias{hideShowIndex}
\alias{hideIndex}
\alias{showIndex}
\title{Hide/Show Index Column}
\usage{
hideIndex(tableId)

showIndex(tableId)
}
\arguments{
\item{tableId}{the id of the table for which the index column is to be hid/shown}
}
\description{
These functions are used to hide or show the index column
(row numbers on far left) in the table.
}
\examples{
if(interactive()) {
 library(shiny)
 library(excelR)
 shinyApp(
   ui = fluidPage(
     excelOutput("table", height = 175),
     actionButton('hideShowButton', 'Hide/Show Index Column')),
   server = function(input, output, session) {
     output$table <- renderExcel(excelTable(data = head(iris)))

     observeEvent(input$hideShowButton,{
       if(input$hideShowButton \%\% 2) # odd number of clicks
         hideIndex("table")
       else # even number of clicks
         showIndex("table")
     })
   }
 )
 }
}
