% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_data.R
\name{getTableData}
\alias{getTableData}
\title{This function is used to get data from excel table}
\usage{
getTableData(tableId)
}
\arguments{
\item{tableId}{the id of the table from which the comment is to be fetched}
}
\description{
This function is used to get data from excel table
}
